WIP:  Bluetooth.
note that debugging lines (marked printk("Fattire: something") or printk("FT") may be in a couple files to help
track down where it's kping at the end of the pairing process.

I have to go but this is for dalingrin to take over for the moment..  the problem seems to be related to

hci_conn_add_sysfs(conn); in net/bluetooth/hci_event.c

The reference isntallation I'm using is at:

http://www.omapzoom.com/?p=kernel/omap.git;a=tree;hb=p-android-omap-2.6.35

Look there to see how things "should" work.