Revert hci_sysfs.c to 2.6.32 version (with modifications)
I guess I jumped the gun by backporting the .35 version.  It built and seemed to work, but kp'd when pairing,
specifically when creating the bluetooth sysfs stuff as it didn't agree with .35.

This is an unbackport to .32 for this file, but still required changes, mostly changing hdev->type to hdev->bus to agree
with the rest of the .35 backport.

Using this file I was able to pair to my mac and transfer a file successfully.